api image/bucketlist/requirements.txt
alembic==0.9.2
aniso8601==1.2.1
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
blinker==1.4
click==6.7
cov-core==1.15.0
coverage==4.4.1
decorator==4.0.11
enum34==1.1.6
Flask==0.12.2
Flask-API==0.7.1
Flask-DebugToolbar==0.10.1
Flask-JWT-Extended==2.4.1
Flask-Migrate==2.0.4
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
Flask-WTF==0.14.2
gunicorn==19.7.1
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
nose==1.3.7
nose2==0.6.5
npm==0.1.1
optional-django==0.1.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.6.1
ptyprocess==0.5.2
pydns==2.3.6
Pygments==2.2.0
PyJWT==1.5.0
python-dateutil==2.6.0
python-decouple==3.0
python-editor==1.0.3
pytz==2017.2
scandir==1.5
simplegeneric==0.8.1
simplejson==3.11.1
simplekv==0.10.0
six==1.10.0
SQLAlchemy==1.1.10
traitlets==4.3.2
validate-email==1.3
wcwidth==0.1.7
Werkzeug==0.12.2
WTForms==2.1
